% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hemoglobin_data}
\alias{hemoglobin_data}
\alias{creatinine_data}
\title{Example values of Hemoglobin and Creatinine}
\format{
\subsection{\code{hemoglobin_data} \code{creatinine_data}}{

A data frame with 1000 rows and 3 columns:
\describe{
\item{age}{age of the patient}
\item{sex}{sex of the patient}
\item{value}{the lab value for the patient, in the default units for the lab}
}
}

An object of class \code{data.frame} with 1000 rows and 3 columns.
}
\usage{
hemoglobin_data

creatinine_data
}
\description{
Example datasets of Hemoglobin and Creatinine values for testing
}
\examples{
head(hemoglobin_data)
head(creatinine_data)
}
\keyword{datasets}
