% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LAB_DETAILS}
\alias{LAB_DETAILS}
\title{Available lab names}
\format{
\subsection{\code{LAB_DETAILS}}{

A data frame with 95 rows and 4 columns:
\describe{
\item{short_name}{Short lab name}
\item{long_name}{Long lab name}
\item{units}{a list column with all the units available for the lab}
\item{default_units}{the default units for the lab}
\item{low_male,high_male,low_female,high_female}{the reference ranges for the lab, taken from the American Board of Internal Medicine. Can be NA if the lab does not have reference ranges.}
\item{clalit_code, ukbb_code}{the codes for the lab in the Clalit and UKBB databases}
}
}
}
\source{
American Board of Internal Medicine. ABIM Laboratory Test Reference Ranges — July 2021. \url{https://www.abim.org/~/media/ABIM\%20Public/Files/pdf/exam/laboratory-reference-ranges.pdf} (2021).
}
\usage{
LAB_DETAILS
}
\description{
Names of the labs available in the package.
}
\examples{
head(LAB_DETAILS)
}
\keyword{datasets}
