% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LAB_INFO}
\alias{LAB_INFO}
\title{Available lab names}
\format{
\subsection{\code{LAB_INFO}}{

A data frame with 65 rows and 4 columns:
\describe{
\item{short_name}{Short lab name}
\item{long_name}{Long lab name}
\item{units}{a list column with all the units available for the lab}
\item{default_units}{the default units for the lab}
}
}
}
\usage{
LAB_INFO
}
\description{
Names of the labs available in the package.
}
\examples{
head(LAB_INFO)
}
\keyword{datasets}
