% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukbb.R
\name{ln_ukbb_labs}
\alias{ln_ukbb_labs}
\title{Get available UKBB labs}
\usage{
ln_ukbb_labs()
}
\value{
A data frame with the available UKBB labs.
}
\description{
Get available UKBB labs
}
\examples{
ln_ukbb_labs()

}
