% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{ln_convert_units}
\alias{ln_convert_units}
\title{Convert values to the default units for the lab}
\usage{
ln_convert_units(values, units, lab)
}
\arguments{
\item{values}{a vector of lab values}

\item{units}{the units of the lab values. See \code{LAB_INFO$units} for a list of available units for each lab. If different values have different units then this should be a vector of the same length as \code{values}.}

\item{lab}{the lab name. See \code{LAB_INFO$short_name} for a list of available labs.}
}
\value{
the values converted to the default units for the lab
}
\description{
Convert values to the default units for the lab
}
\examples{

# emulate a dataset with different units

hemoglobin_diff_units <- hemoglobin_data

# first 500 values will be in mg/ML
hemoglobin_diff_units$value[1:500] <- hemoglobin_diff_units$value[1:500] * 10

# last 500 values will be in mmol/L
hemoglobin_diff_units$value[501:1000] <- hemoglobin_diff_units$value[501:1000] / 1.61


converted <- ln_convert_units(
    hemoglobin_diff_units$value,
    c(rep("mg/mL", 500), rep("mmol/L", 500)),
    "Hemoglobin"
)

head(converted)
head(hemoglobin_data$value)
}
