% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-quantiles.R
\name{ln_quantile_value}
\alias{ln_quantile_value}
\title{Compute the lab value for a given quantile}
\usage{
ln_quantile_value(quantiles, age, sex, lab)
}
\arguments{
\item{quantiles}{a vector of quantiles (in the range 0-1) to compute the lab value for.}

\item{age}{a vector of ages between (20-99). Can be a single value if all values are the same age.}

\item{sex}{a vector of either "male" or "female". Can be a single value if all values are the same sex.}

\item{lab}{The lab name.}
}
\value{
a data frame which contains the values for each combination of quantile, age and sex.
The data frame has the the following columns:

\itemize{
\item{age: }{age in years}
\item{sex: }{"male" or "female"}
\item{quantile: }{he quantile}
\item{value: }{the lab value}
\item{unit: }{the lab unit}
\item{lab: }{the lab name}
}
}
\description{
This function computes the lab value for given quantiles in the default units for the lab. Default units can be found using the function \code{ln_lab_default_units}. In cases where the original data had only a single value, NA is returned.
}
\examples{
ln_quantile_value(c(0, 0.03, 0.5, 0.97, 1), 50, "male", "WBC")

ln_quantile_value(
    c(0, 0.03, 0.1, 0.4, 0.5, 0.6, 0.9, 0.97, 1),
    c(50, 60),
    c("male", "female"),
    "Glucose"
)

}
